Implementando componente de Tabela de Alimento Consolidado
